requirements.txt
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
bert-score==0.3.13
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
compressed-tensors==0.8.0
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
evaluate==0.4.3
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.26.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
natsort==8.4.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge_score==0.1.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece @ file:///Users/runner/miniforge3/conda-bld/sentencepiece-split_1730804082527/work/python
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.18.0
